Moved methods to the abstract super class
